Link accounts to users

+++ src/main/java/com/alshubaily/fintech/tiger_ledger_service/model/account/response/CreateAccountResponse.java
@@ -1,3 +1,5 @@
 package com.alshubaily.fintech.tiger_ledger_service.model.account.response;
 
-public record CreateAccountResponse(long accountId) { }
+import java.math.BigInteger;
+
+public record CreateAccountResponse(BigInteger accountId) { }
